ecom/requirements.txt
django==2.2
djangorestframework==3.8.2
django-cors-headers==2.4.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.6
djangorestframework-jwt==1.11.0